+++ MekHQ/src/mekhq/campaign/personnel/death/RandomDeath.java
@@ -54,9 +54,8 @@ import megamek.common.enums.Gender;
 import megamek.common.util.weightedMaps.WeightedDoubleMap;
 import megamek.logging.MMLogger;
 import mekhq.MHQConstants;
-import mekhq.MekHQ;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.Injury;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.personnel.enums.AgeGroup;
