+++ MekHQ/src/mekhq/campaign/personnel/skills/RandomSkillPreferences.java
@@ -40,7 +40,7 @@ import java.util.Map.Entry;
 
 import megamek.Version;
 import megamek.logging.MMLogger;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.enums.PersonnelRole;
 import mekhq.utilities.MHQXMLUtility;
 import org.w3c.dom.Node;
