+++ MekHQ/src/mekhq/gui/campaignOptions/contents/RepairAndMaintenanceTab.java
@@ -43,7 +43,7 @@ import javax.swing.JPanel;
 import javax.swing.JSpinner;
 
 import megamek.common.annotations.Nullable;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.gui.campaignOptions.components.CampaignOptionsCheckBox;
 import mekhq.gui.campaignOptions.components.CampaignOptionsGridBagConstraints;
 import mekhq.gui.campaignOptions.components.CampaignOptionsHeaderPanel;
